\name{file.swap}
\alias{file.swap}
\title{Swap names of two files (by renaming first to a temporary file)}
\usage{
  file.swap(f1, f2)
}
\arguments{
  \item{f1,f2}{Paths to files}
}
\value{
  logical indicating success
}
\description{
  Swap names of two files (by renaming first to a temporary
  file)
}
\author{
  jefferis
}
\seealso{
  \code{\link{file.rename}}
}

