% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{make_nl_from_parts}
\alias{make_nl_from_parts}
\title{Make a neuronlist object from two separate files}
\usage{
make_nl_from_parts(datapath, dfpath, extdata_package = NULL, ...)
}
\arguments{
\item{datapath}{Path to the data object}

\item{dfpath}{Path to the data.frame object}

\item{extdata_package}{Character vector naming a package whose extdata
directory will be sought (with \code{\link{find_extdata}}) and prepended to
the two input paths.}

\item{...}{Additional arguments passd to \code{\link{find_extdata}}}
}
\value{
a \code{neuronlist} object
}
\description{
Make a neuronlist object from two separate files
}
