% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fsutils.r
\name{nat.utils-defunct}
\alias{file.hardlink}
\alias{nat.utils-defunct}
\title{nat.utils: Defunct functions}
\usage{
file.hardlink(from, to)
}
\arguments{
\item{from}{Source file}

\item{to}{(New) target hardlink file to create}
}
\description{
These functions have been retired from \code{nat.utils}
}
\seealso{
\code{\link{file.link}}
}

