% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{nat.utils}
\alias{nat.utils}
\alias{nat.utils-package}
\title{nat.utils: File System Utility Functions for NeuroAnatomy Toolbox}
\description{
Utility functions to support the NeuroAnatomy Toolbox (nat).
Includes functions to provide a basic make style system to update files
based on timestamp information, file locking and other convenience functions
for working with the filesystem
}
\seealso{
nat
}

