% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsutils.r
\name{touch}
\alias{touch}
\title{Use unix touch utility to change file's timestamp}
\usage{
touch(
  file,
  time,
  reference,
  timestoupdate = c("access", "modification"),
  Create = TRUE
)
}
\arguments{
\item{file}{Path to file to modify}

\item{time}{Absolute time in POSIXct format}

\item{reference}{Path to a reference file}

\item{timestoupdate}{"access" or "modification" (default both)}

\item{Create}{Logical indicating whether to create file (default TRUE)}
}
\value{
TRUE or FALSE according to success
}
\description{
If neither a time or a reference file is provided then the current time is 
used. If the file does not already exist, it is created unless Create=FALSE.
}
\author{
jefferis
}
