% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{find_extdata}
\alias{find_extdata}
\title{Construct paths to files in the extdata folder of a package}
\usage{
find_extdata(..., package = NULL, firstpath = file.path("inst", "extdata"))
}
\arguments{
\item{...}{components of the path (eventually appended to location of
\code{extdata})}

\item{package}{The package to search}

\item{firstpath}{A hint for the first place to look}
}
\value{
A character vector containing the constructed path
}
\description{
Construct paths to files in the extdata folder of a package
}
\details{
\code{inst/extdata} is the conventional place to store data this is
  not managed directly by the standard R package mechanisms. Unfortunately
  its location changes at different stages of the package build/load process,
  since in the final package all folders under \code{inst} are installed
  directly under the package root.
}
\examples{
find_extdata(package='nat.utils')
}
