% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziputils.r
\name{zipinfo}
\alias{zipinfo}
\title{Return information about a zip archive using system unzip command}
\usage{
zipinfo(f)
}
\arguments{
\item{f}{Path to one (or more) files}
}
\value{
dataframe of information
}
\description{
Return information about a zip archive using system unzip command
}
\details{
Uses system unzip command.
}
\seealso{
\code{\link{zip}}

Other ziputils: 
\code{\link{zipok}()}
}
\author{
jefferis
}
\concept{ziputils}
